tool/ceph_dedup: split ceph_dedup_tool into ceph_dedup_tool and ceph_dedup_daemon
