+++ src/tools/CMakeLists.txt
@@ -122,17 +122,6 @@ add_executable(ceph-authtool ${ceph_authtool_srcs})
 target_link_libraries(ceph-authtool global ${EXTRALIBS} ${CRYPTO_LIBS})
 install(TARGETS ceph-authtool DESTINATION bin)
 
-if(WITH_TESTS)
-set(ceph_dedup_tool_srcs ceph_dedup_tool.cc)
-add_executable(ceph-dedup-tool ${ceph_dedup_tool_srcs})
-target_link_libraries(ceph-dedup-tool
-  librados
-  global
-  cls_cas_client
-  cls_cas_internal)
-install(TARGETS ceph-dedup-tool DESTINATION bin)
-endif(WITH_TESTS)
-
 if(WITH_CEPHFS)
   add_subdirectory(cephfs)
   add_subdirectory(cephfs_mirror)
@@ -155,3 +144,4 @@ endif(WITH_RBD)
 add_subdirectory(immutable_object_cache)
 add_subdirectory(ceph-dencoder)
 add_subdirectory(erasure-code)
+add_subdirectory(ceph_dedup)
